create / update website
Creation of the Angular project with creation of the different components.
A Topbar component which contains the html code of the top bar with the connection profile and the language change (not done yet). A second component "project-manager" allows you to manage projects (create, delete, access a project, etc.). And finally, a last "project-editor" component to edit a project, add entities (image/sound/video), delete them.
This entire part is the frontend on the administrator/organizer side.