update Website
Wave 3.5 Added the frontend for the user by setting up the routing system in Angular. Creation of the Admin component (to manage the view and the components for the administrator view) and User (for the user view). Added some test methods which allows you to finish the get api tests of the Project Manager component. On the Project Editor side, implementation of get / post method to call the API. After changes to the api (openAPI), modification of some classes.